cap nhat

+++ fruitapp/src/main/java/com/example/fruitapp/repository/UserRepository.java
@@ -1,12 +1,16 @@
 package com.example.fruitapp.repository;
 
 import com.example.fruitapp.entity.UserEntity;
+
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository
-        extends JpaRepository<UserEntity, Integer> {
+        extends JpaRepository<UserEntity, Long> {
 
-    Optional<UserEntity> findByUsername(String username);
+    Optional<UserEntity> findByUsername(
+            String username
+    );
 }
+
